requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dateparser